+++ test/cross_signing/install_faketime.sh
@@ -1,0 +1,1 @@
+../common_scenarios/install_faketime.sh